Propensity score matched elective laparoscopic cholecystectomies from NSQIP Registry

A dataset containing 48,240 records of elective laparoscopic cholecystectomies from the NSQIP registry. This data was derived from a larger dataset, nsqipBileSpill::df, by propensity score matching for assignment to the variable spill against various pre-operative risk factors.

df_match

Format

A data frame with 48,240 rows and 56 variables:

sex

Logical, TRUE for male, FALSE for female

age

Age in years

bmi

BMI of patient in kg/m^2

diabetes

Diabetes mellitus with oral agents or insulin

insulin

Insulin dependent diabetes mellitus

smoke

Current smoker within 1 year

dyspnea

Presence of dypsnea

preopindependent

Functionally independent prior to surgery

hxcopd

History of severe COPD

ascites

Ascites within 30 days prior to surgery

hxchf

Congestive heart failure within 30 days prior to surgery

hypermed

Hyperetnsion requiring medication

dialysis

Currently requiring or on dialysis

discancr

Disseminated cancer

steroid

Steroid/immunosuppressant use for a chronic condition

wtloss

>10% weight loss of body weight within 6 months prior to surgery

bleeddis

Nleeding disorders

spill

Presumptive bile spill, based on wound classification

optime

Total operation time in minutes

supinfec

Superficial incisional surgical site infection

wndinfd

Deep incisional surgical site infection

orgspcssi

Organ-space surgical site infection

mort

Death within 30 days of surgery

tothlos

Length of total hospital stay in days

dindo_min

Minor complication (Dindo-Clavien 1 or 2)

dindo_maj

Major complication (Dindo-Clavien 3 or more)

ioc

Intraoperative cholangiogram performed

distance

Estimated distance measure from propenstiy score match

weights

The weights assigned to each unit in the matching process. Unmatched units have weights equal to 0. Matched treated units have weight 1. Each matched control unit has weight proportional to the number of treatment units to which it was matched, and the sum of the control weights is equal to the number of uniquely matched control units

Source

https://www.facs.org/quality-programs/acs-nsqip